+++ torch/_C/_autograd.pyi
@@ -85,7 +85,7 @@ def _disable_profiler() -> _ProfilerResult: ...
 def _profiler_enabled() -> bool: ...
 def _add_metadata_json(key: str, value: str) -> None: ...
 def kineto_available() -> bool: ...
-def _supported_kineto_activities() -> Set[ProfilerActivity]: ...
+def _supported_activities() -> Set[ProfilerActivity]: ...
 def _enable_record_function(enable: bool) -> None: ...
 def _set_empty_test_observer(is_global: bool, sampling_prob: float) -> None: ...
 def _register_default_hooks(pack_hook: Callable, unpack_hook: Callable) -> None: ...
