+++ torch/csrc/autograd/init.cpp
@@ -115,10 +115,11 @@ PyObject* THPAutograd_initExtension(PyObject* _unused, PyObject *unused) {
       .value("Vulkan", c10::DeviceType::Vulkan)
       .value("Metal", c10::DeviceType::Metal);
 
-#ifdef USE_KINETO
   py::class_<KinetoEvent>(m, "_KinetoEvent")
       // name of the event
-      .def("name", &KinetoEvent::name)
+      .def("name", [](const KinetoEvent& e) {
+        return e.name();
+      })
       // PyTorch thread id of the start callback
       .def("start_thread_id", [](const KinetoEvent& e) {
         return e.startThreadId();
@@ -138,9 +139,13 @@ PyObject* THPAutograd_initExtension(PyObject* _unused, PyObject *unused) {
         return e.sequenceNr();
       })
       // absolute start time (since unix epoch) in us
-      .def("start_us", &KinetoEvent::startUs)
+      .def("start_us", [](const KinetoEvent& e) {
+        return e.startUs();
+      })
       // duration in us
-      .def("duration_us", &KinetoEvent::durationUs)
+      .def("duration_us", [](const KinetoEvent& e) {
+        return e.durationUs();
+      })
       // used for correlation between high-level PyTorch events
       // and low-level device events
       .def("correlation_id", [](const KinetoEvent& e) {
@@ -179,13 +184,17 @@ PyObject* THPAutograd_initExtension(PyObject* _unused, PyObject *unused) {
         return e.deviceIndex();
       })
       // for CUDA - stream id, for CPU - start thread id
-      .def("device_resource_id", &KinetoEvent::deviceResourceId)
+      .def("device_resource_id", [](const KinetoEvent& e) {
+        return e.deviceResourceId();
+      })
       // device type
       .def("device_type", [](const KinetoEvent& e) {
         return e.deviceType();
       })
       // correlation id of a linked event
-      .def("linked_correlation_id", &KinetoEvent::linkedCorrelationId)
+      .def("linked_correlation_id", [](const KinetoEvent& e) {
+        return e.linkedCorrelationId();
+      })
       // compute flops
       .def("flops", [](const KinetoEvent& e) {
         return e.flops();
@@ -202,12 +211,14 @@ PyObject* THPAutograd_initExtension(PyObject* _unused, PyObject *unused) {
   py::class_<ProfilerResult>(m, "_ProfilerResult")
     .def("trace_start_us", &ProfilerResult::trace_start_us)
     .def("events", &ProfilerResult::events)
-    .def("save", &ProfilerResult::save);
+#ifdef USE_KINETO
+    .def("save", &ProfilerResult::save)
+#endif // USE_KINETO
+    ;
 
   m.def("_enable_profiler", enableProfiler);
   m.def("_disable_profiler", disableProfiler);
   m.def("_prepare_profiler", prepareProfiler);
-#endif
 
   m.def("_add_metadata_json", [](const std::string& key, const std::string& value) {
 #ifdef USE_KINETO
@@ -215,7 +226,7 @@ PyObject* THPAutograd_initExtension(PyObject* _unused, PyObject *unused) {
 #else
       LOG(WARNING) << "Adding profiling metadata requires using "
                    << "torch.profiler with Kineto support (USE_KINETO=1)";
-#endif
+#endif // USE_KINETO
   });
 
   m.def("kineto_available", []() {
@@ -226,15 +237,12 @@ PyObject* THPAutograd_initExtension(PyObject* _unused, PyObject *unused) {
 #endif
   });
 
-  m.def("_supported_kineto_activities", []() {
-    std::set<ActivityType> activities;
-#ifdef USE_KINETO
-    activities.insert(ActivityType::CPU);
-#ifndef LIBKINETO_NOCUPTI
+  m.def("_supported_activities", []() {
+    std::set<ActivityType> activities {ActivityType::CPU};
+#if defined(USE_KINETO) && !defined(LIBKINETO_NOCUPTI)
     if (at::getNumGPUs() > 0 && !at::hasHIP()) {
       activities.insert(ActivityType::CUDA);
     }
-#endif
 #endif
     return activities;
   });
