+++ torch/csrc/autograd/profiler_kineto.h
@@ -26,8 +26,6 @@ enum class C10_API_ENUM ActivityType {
   NUM_KINETO_ACTIVITIES, // must be the last one
 };
 
-#ifdef USE_KINETO
-
 // NOLINTNEXTLINE(cppcoreguidelines-pro-type-member-init)
 struct KinetoObserverContext : public at::ObserverContext {
   int64_t startUs;
@@ -51,16 +49,36 @@ struct TORCH_API KinetoEvent {
     return start_thread_id_;
   }
 
+  KinetoEvent& startThreadId(uint64_t start_thread_id) {
+    start_thread_id_ = start_thread_id;
+    return *this;
+  }
+
   uint64_t endThreadId() const {
     return end_thread_id_;
+  }
+
+  KinetoEvent& endThreadId(uint64_t end_thread_id) {
+    end_thread_id_ = end_thread_id;
+    return *this;
   }
 
   uint8_t activityType() const {
     return activity_type_;
   }
 
+  KinetoEvent& activityType(uint8_t activity_type) {
+    activity_type_ = activity_type;
+    return *this;
+  }
+
   uint64_t fwdThreadId() const {
     return fwd_thread_id_;
+  }
+
+  KinetoEvent& fwdThreadId(uint64_t fwd_thread_id) {
+    fwd_thread_id_ = fwd_thread_id;
+    return *this;
   }
 
   bool hasShapes() const {
@@ -71,6 +89,11 @@ struct TORCH_API KinetoEvent {
     return *shapes_;
   }
 
+  KinetoEvent& shapes(const std::vector<std::vector<int64_t>>& shapes) {
+    shapes_ = shapes;
+    return *this;
+  }
+
   bool hasTypes() const {
     return dtypes_ != c10::nullopt;
   }
@@ -79,12 +102,27 @@ struct TORCH_API KinetoEvent {
     return *dtypes_;
   }
 
+  KinetoEvent& dtypes(const std::vector<std::string>& dtypes) {
+    dtypes_ = dtypes;
+    return *this;
+  }
+
   uint64_t flops() const {
     return flops_;
   }
 
+  KinetoEvent& flops(uint64_t flops) {
+    flops_ = flops;
+    return *this;
+  }
+
   int64_t sequenceNr() const {
     return sequence_nr_;
+  }
+
+  KinetoEvent& sequenceNr(int64_t sequence_nr) {
+    sequence_nr_ = sequence_nr;
+    return *this;
   }
 
   bool hasStack() const {
@@ -95,52 +133,26 @@ struct TORCH_API KinetoEvent {
     return *stack_;
   }
 
-  uint8_t scope() const {
-    return scope_;
-  }
-
-  KinetoEvent& startThreadId(uint64_t start_thread_id) {
-    start_thread_id_ = start_thread_id;
-    return *this;
-  }
-
-  KinetoEvent& endThreadId(uint64_t end_thread_id) {
-    end_thread_id_ = end_thread_id;
-    return *this;
-  }
-
-  KinetoEvent& fwdThreadId(uint64_t fwd_thread_id) {
-    fwd_thread_id_ = fwd_thread_id;
-    return *this;
-  }
-
-  KinetoEvent& shapes(const std::vector<std::vector<int64_t>>& shapes) {
-    shapes_ = shapes;
-    return *this;
-  }
-
-  KinetoEvent& dtypes(const std::vector<std::string>& dtypes) {
-    dtypes_ = dtypes;
-    return *this;
-  }
-
-  KinetoEvent& flops(uint64_t flops) {
-    flops_ = flops;
-    return *this;
-  }
-
-  KinetoEvent& sequenceNr(int64_t sequence_nr) {
-    sequence_nr_ = sequence_nr;
-    return *this;
-  }
-
   KinetoEvent& stack(const std::vector<std::string>& st) {
     stack_ = st;
     return *this;
   }
 
+  uint8_t scope() const {
+    return scope_;
+  }
+
   KinetoEvent& scope(uint8_t scope) {
     scope_ = scope;
+    return *this;
+  }
+
+  std::string name() const {
+    return name_;
+  }
+
+  KinetoEvent& name(const std::string& evt_name) {
+    name_ = evt_name;
     return *this;
   }
 
@@ -149,9 +161,17 @@ struct TORCH_API KinetoEvent {
     return *this;
   }
 
+  c10::DeviceType deviceType() const {
+    return (c10::DeviceType)device_type_;
+  }
+
   KinetoEvent& deviceType(c10::DeviceType device_type) {
     device_type_ = (int8_t)device_type;
     return *this;
+  }
+
+  uint8_t deviceIndex() const {
+    return device_index_;
   }
 
   KinetoEvent& deviceIndex(uint8_t device_index) {
@@ -159,33 +179,35 @@ struct TORCH_API KinetoEvent {
     return *this;
   }
 
+  int64_t nBytes() const {
+    return nbytes_;
+  }
+
   KinetoEvent& nBytes(int64_t nbytes) {
     nbytes_ = nbytes;
     return *this;
-  }
-
-  // Kineto fields
-
-  KinetoEvent& activity(const libkineto::TraceActivity& activity);
-
-  std::string name() const {
-    return name_;
-  }
-
-  bool isAsync() const {
-    return is_async_;
-  }
-
-  uint8_t deviceIndex() const {
-    return device_index_;
   }
 
   uint64_t startUs() const {
     return start_us_;
   }
 
+  KinetoEvent& startUs(uint64_t start_us) {
+    start_us_ = start_us;
+    return *this;
+  }
+
   uint64_t durationUs() const {
     return duration_us_;
+  }
+
+  KinetoEvent& durationUs(uint64_t duration_us) {
+    duration_us_ = duration_us;
+    return *this;
+  }
+
+  bool isAsync() const {
+    return is_async_;
   }
 
   uint64_t correlationId() const {
@@ -201,15 +223,19 @@ struct TORCH_API KinetoEvent {
     return linked_correlation_id_;
   }
 
+  KinetoEvent& linkedCorrelationId(uint64_t linked_correlation_id) {
+    linked_correlation_id_ = linked_correlation_id;
+    return *this;
+  }
+
   int64_t deviceResourceId() const {
     return device_resource_id_;
   }
 
-  int64_t nBytes() const {
-    return nbytes_;
+  KinetoEvent& deviceResourceId(int64_t device_resource_id) {
+    device_resource_id_ = device_resource_id;
+    return *this;
   }
-
-  c10::DeviceType deviceType() const;
 
   int64_t cudaElapsedUs() const;
 
@@ -227,7 +253,7 @@ struct TORCH_API KinetoEvent {
 
   std::string name_;
   uint8_t device_index_ = 0;
-  int8_t device_type_ = -1;
+  int8_t device_type_ = 0;
   uint64_t start_us_ = 0;
   uint64_t duration_us_ = 0;
   uint64_t correlation_id_ = 0;
@@ -245,10 +271,14 @@ struct TORCH_API KinetoEvent {
 // memory allocation events)
 struct TORCH_API ProfilerResult {
   ProfilerResult();
+#ifdef USE_KINETO
   ProfilerResult(
       uint64_t start_time,
       std::vector<KinetoEvent> events,
       std::unique_ptr<libkineto::ActivityTraceInterface> trace);
+#else
+  ProfilerResult(std::vector<KinetoEvent> events);
+#endif // USE_KINETO
   ~ProfilerResult();
 
   uint64_t trace_start_us() const {
@@ -259,13 +289,17 @@ struct TORCH_API ProfilerResult {
     return events_;
   }
 
+#ifdef USE_KINETO
   void save(const std::string& path);
+#endif // USE_KINETO
 
  private:
-  bool saved_ = false;
   uint64_t trace_start_us_ = 0;
   std::vector<KinetoEvent> events_;
+#ifdef USE_KINETO
   std::unique_ptr<libkineto::ActivityTraceInterface> trace_;
+  bool saved_ = false;
+#endif // USE_KINETO
 };
 
 TORCH_API void enableProfiler(
@@ -280,7 +314,6 @@ TORCH_API void prepareProfiler(
 
 TORCH_API void addMetadataJson(
     const std::string& key, const std::string& value);
-#endif // USE_KINETO
 
 } // namespace profiler
 }} // namespace torch::autograd
